# platform/resource_managers/swagger_server/swagger/swagger-old.yaml
...
swagger: "2.0"
info:
  description: "Prototype of the RM API "
  version: "1.0"
  title: "RM Prototype"
host: "virtserver.swaggerhub.com"
basePath: "/jvandebon/rm_prototype/1.0"
...
schemes:
- "https"
- "http"
paths:
...
  /tasks/config:
    get:
      tags:
      - "tasks"
      summary: "Returns optimal configuration"
      operationId: "optimal_config"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "task"
        description: "Function to be executed with input parameters and objectives"
        required: true
        schema:
          $ref: "#/definitions/task_1"
      responses:
        200:
          description: "Successful config determination"
          schema:
            type: "object"
            properties:
              impl:
                $ref: "#/definitions/inline_response_200_2_impl"
              configuration:
                $ref: "#/definitions/inline_response_200_2_configuration"
      x-swagger-router-controller: "swagger_server.controllers.tasks_controller"
...
definitions:
...
  task_1:
    type: "object"
    properties:
      taskName:
        type: "string"
        example: "adp"
      problemSize:
        type: "integer"
        example: 1000
  inline_response_200_2_configuration:
    properties:
      name:
        type: "string"
        example: "4T"
      description:
        type: "string"
        example: "4 CPU threads"
      n:
        type: "integer"
        example: 4
    example:
      name: "4T"
      description: "4 CPU threads"
      n: 4
  inline_response_200_2_impl:
    properties:
      taskName:
        type: "string"
        example: "adp"
      implName:
        type: "string"
        example: "cpu_compute"
      implRM:
        type: "string"
        example: "hnode0.cpu0"
    example:
      implName: "cpu_compute"
      taskName: "adp"
      implRM: "hnode0.cpu0"
...